Create base UDP and Serial classes to be inherited
