+++ ledfx/devices/__init__.py
@@ -7,6 +7,10 @@ import numpy as np
 import voluptuous as vol
 import zeroconf
 
+import socket
+import serial
+import serial.tools.list_ports
+
 from ledfx.config import save_config
 from ledfx.events import DeviceUpdateEvent, Event
 from ledfx.utils import (
@@ -38,6 +42,14 @@ class Device(BaseRegistry):
     def CONFIG_SCHEMA():
         return vol.Schema(
             {
+                vol.Required(
+                    "name", description="Friendly name for the device"
+                ): str,
+                vol.Required(
+                    "pixel_count",
+                    description="Number of individual pixels",
+                    default=1,
+                ): vol.All(vol.Coerce(int), vol.Range(min=1)),
                 vol.Optional(
                     "icon_name",
                     description="https://material-ui.com/components/material-icons/",
@@ -67,6 +79,7 @@ class Device(BaseRegistry):
         self._segments = []
         self._pixels = None
         self._silence_start = None
+        self._device_type = ""
 
     def __del__(self):
         if self._active:
@@ -335,6 +348,80 @@ class NetworkedDevice(Device):
             return
         else:
             return self._destination
+
+
+@BaseRegistry.no_registration
+class UDPDevice(NetworkedDevice):
+
+    CONFIG_SCHEMA = vol.Schema(
+        {
+            vol.Required(
+                "port",
+                description="Port for the UDP device",
+            ): vol.All(vol.Coerce(int), vol.Range(min=1, max=65535)),
+        }
+    )
+
+    def activate(self):
+        self._sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+        _LOGGER.info(f"{self._device_type} sender for {self.config['name']} started.")
+        super().activate()
+
+    def deactivate(self):
+        super().deactivate()
+        _LOGGER.info(f"{self._device_type} sender for {self.config['name']} stopped.")
+        self._sock = None
+
+
+class AvailableCOMPorts:
+    ports = serial.tools.list_ports.comports()
+
+    available_ports = []
+
+    for p in ports:
+        available_ports.append(p.device)
+
+@BaseRegistry.no_registration
+class SerialDevice(Device):
+
+    CONFIG_SCHEMA = vol.Schema(
+        {
+            vol.Required(
+                "com_port",
+                description="COM port for Adalight compatible device",
+            ): vol.In(list(AvailableCOMPorts.available_ports)),
+            vol.Required(
+                "baudrate", description="baudrate", default=500000
+            ): vol.All(vol.Coerce(int), vol.Range(min=115200)),
+        }
+    )
+
+    def __init__(self, ledfx, config):
+        super().__init__(ledfx, config)
+        self.serial = None
+        self.baudrate = self._config["baudrate"]
+        self.com_port = self._config["com_port"]
+
+    def activate(self):
+        try:
+            if self.serial and self.serial.isOpen:
+                return
+
+            self.serial = serial.Serial(self.com_port, self.baudrate)
+            if self.serial.isOpen:
+                super().activate()
+
+        except serial.SerialException:
+            _LOGGER.critical(
+                "Serial Error: Please ensure your device is connected, functioning and the correct COM port is selected."
+            )
+            # Todo: Trigger the UI to refresh after the clear effect call. Currently it still shows as active.
+            self.deactivate()
+
+    def deactivate(self):
+        super().deactivate()
+        if self.serial:
+            self.serial.close()
 
 
 class Devices(RegistryLoader):
